refactor(faq): FAQ 도메인 추상화 — 서브컴포넌트 추출 및 이름 정리
- AsyncBoundary 공용 컴포넌트 생성 (common/boundary/)
- FaqSection에서 FaqInput, FaqButton, FaqItem 서브컴포넌트 추출 (domain/faq/ui/)
- FaqSection 로직을 useFaqList 커스텀 훅으로 분리 (domain/faq/hooks/)
- FaqChat 내부 함수명 FaqDropdown → FaqChat으로 변경
