+++ apps/web/src/domain/faq/FaqChat.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 
 import channelService from '@/ChannelService';
 
-function FaqDropdown() {
+function FaqChat() {
   return (
     <div className="bg-neutral-95 mx-auto flex flex-col items-center gap-3 rounded-md px-8 py-4 md:w-full md:max-w-[800px] md:flex-row md:items-center md:justify-between">
       <span className="text-xsmall14 text-neutral-35 md:text-small20 font-semibold">
@@ -18,4 +18,4 @@ function FaqDropdown() {
   );
 }
 
-export default memo(FaqDropdown);
+export default memo(FaqChat);
